bsp/xilinx_zynq: Support configuraton of memory map. Remove SMP variants.
Generate a linker command file from configure letting the user override the
defaults to suite their custom needs. Refer to configure.ac for the details.

Remove the SMP variants and let --enable-smp control if a BSP is built for
SMP.

Make USE_FAST_IDLE 1 only for the realview qemu BSP.