score: Add _CPU_Get_current_per_CPU_control()
Add optional method _CPU_Get_current_per_CPU_control() to obtain the
per-CPU control of the current processor.

This is optional.  Not every CPU port needs this.  It is only an
optional optimization variant.  In case this macro is undefined, the
default implementation using the current processor index will be used.