score: Fix _Thread_Start_multitasking() on SMP
The _CPU_Context_Restart_self() implementations usually assume that self
context is executing.

FIXME: We have a race condition in _Thread_Start_multitasking() in case
another thread already performed scheduler operations and moved the heir
thread to another processor.  The time frame for this is likely too
small to be practically relevant.