Remove libcpu/mips/clock - unused code
This code was only referenced for r46xx CPU models. There are
no BSPs currently in the tree with this CPU model. The r46xx
was the original target for the MIPS port and it is likely that
the code has not been used on actual hardware in many years.

All MIPS BSPs successfully built after this was removed.